Handle the prototype attribute during sub creation
This is necessary so that the prototype attribute can
serve as a drop-in replacement for an actual prototype.

The attribute is shifted out of the attribute list before
the attributes handler is ever called, and before the interpreter
starts to handle the prototype.  This necessitates a new warning,
since declaring a prototype and the prototype attribute at the
same time leads to the attribute replacing the traditional
declaration in all senses that matter.